% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rminuit2-package.R
\docType{package}
\name{rminuit2-package}
\alias{rminuit2-package}
\title{rminuit2: Function Minimization with Minuit2}
\description{
R interface for Minuit2 function minimization library (https://root.cern.ch/doc/master/Minuit2Page.html), which is included in the package.
}
\author{
\strong{Maintainer}: Alberto Lusiani \email{alusiani@gmail.com}

}
