% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rminuit2.R
\name{rminuit2}
\alias{rminuit2}
\title{Function Minimization with Minuit2}
\usage{
rminuit2(fn, par, err = NULL, lower = NULL, upper = NULL, fix = NULL,
  opt = "h", envir = NULL, ..., maxcalls = 0L, nsigma = 1)
}
\arguments{
\item{fn}{The function to be minimized, with first argument a numeric
vector of the parameters to be optimized. The function should be either an R function
or an external pointer to a C++ function compiled using the
\code{inline} interface. For more info about implementing the objective
function as compiled C++ code, see the lbfgs package vignette}

\item{par}{numeric vector, initial values of the function parameters}

\item{err}{numeric vector, expected uncertainty of function parameters}

\item{lower}{numeric vector, lower bounds for the parameters (default none)}

\item{upper}{numeric vector, upper bounds for the parameters (default none)}

\item{fix}{boolean vector, each TRUE element fixes the corresponding parameter}

\item{opt}{string, pass fit options, default "h" (compute HESSE errors)
\itemize{
\item{v}{Verbose mode (not yet implemented)}
\item{h}{Run Hesse to estimate errors}
\item{m}{Get Minos errors}
\item{0}{Run Migrad with strategy 0}
\item{1}{Run Migrad with strategy 1}
\item{2}{Run Migrad with strategy 2}
\item{no 123}{Run Migrad with strategy 1 and if fails 2}
}}

\item{envir}{An R environment containing all extra arguments to be passed
to the objective function, which must be matched exactly. If the objective
function is implemented in C++, extra arguments must be passed using this option,
rather than the \code{...} construct. If the functions are implemented in R, extra
arguments should be passed to them using the \code{...} construct instead.}

\item{...}{construct. If the functions are implemented in R, extra arguments
should be passed to them using the \code{...} construct instead.}

\item{maxcalls}{integer, maximum number of calls, defaults to \code{0} (no limit).}

\item{nsigma}{numeric, number of standard deviations for Minos errors}
}
\value{
A list with the following components:
 \itemize{
   \item{value}{The minimized value of the objective function.}
   \item{par}{A numerical array with the parameters that minimize the function.}
   \item{par_err}{A numerical array with the estimeted parameters' uncertainties.}
   \item{par_cov}{A matrix with the estimated covariance matrix of the parameters.}
   \item{convergence}{An integer code. Zero indicates that convergence was reached without issues. Negative values indicate errors in the execution of the L-BFGS routine.}
   \item{message}{A character object detailing execution errors. This component is only returned if the convergence code is different form zero.}
 }
}
\description{
Performs function minimization using
\href{https://project-mathlibs.web.cern.ch/project-mathlibs/sw/Minuit2/html/index.html}{Minuit2}
}
\examples{
# Rosenbrock Banana function

rosenbrock <- function(x) {
    x1 <- x[1]
    x2 <- x[2]
    100 * (x2 - x1 * x1)^2 + (1 - x1)^2
}

output <- rminuit2(rosenbrock, c(-1.2, 1))

}
\author{
Alberto Lusiani, \email{alusiani@gmail.com}
}
\keyword{fitting}
\keyword{minimization}
\keyword{optimization}
